requirements.txt
appdirs==1.4.3
argh==0.26.2
enum34==1.1.6
Markdown==2.6.8
markslide==0.1
packaging==16.8
pathlib==1.0.1
pathtools==0.1.2
pdfkit==0.6.1
pyparsing==2.2.0
pyScss==1.3.5
PyYAML==3.12
six==1.10.0
watchdog==0.8.3
